test(snapshot): Migrate more tests
